Changed id type from serializable to Long

+++ src/main/java/ar/edu/unq/dapp/c2a/webServices/controllers/OrdersController.java
@@ -23,7 +23,7 @@ public class OrdersController {
     public @ResponseBody
     Collection<OrderDTO> getOrders() {
         // TODO: replace with  actual implementation
-        Integer clientId = 1;
+        long clientId = 1l;
 
         return orderService.getClientOrders(clientId);
     }
